requirements.txt
asgiref==3.8.1
Django==5.2.2
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
Markdown==3.8
psycopg2-binary==2.9.10
sqlparse==0.5.3
gunicorn